refactor(release): Generalize git log prompt for flexibility
The prompt for the `release/prepare` command has been updated to be less prescriptive about how to retrieve the git commit history.

Previously, the prompt included a specific `git log` command with a fixed format. This has been replaced with a more general instruction to get the commit history, only specifying that the `--no-pager` option must be used if the `git log` command is chosen.

This change provides the language model with more flexibility to determine the most appropriate command and format for generating the changelog, improving the robustness of the release preparation process.